Routing Completed (Rebased some stuff)
